trie: fix dirty storage size error
